api/pyproject.toml
[project]
name = "balancenote-api"
version = "0.1.0"
description = "BalanceNote API — FastAPI + LangGraph Self-RAG"
requires-python = ">=3.13"
dependencies = [
    # --- Web framework ---
    "fastapi[standard]>=0.115",
    # --- LangChain / LangGraph ---
    "langgraph>=1.1.9",
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    # Story 3.3 — AsyncPostgresSaver API는 3.x에서 안정 (lock 3.0.5 정합). 외주 인수
    # readiness — 2.x는 deprecated 신호.
    "langgraph-checkpoint-postgres>=3.0",
    "langsmith>=0.3",
    # --- DB / ORM ---
    "sqlalchemy[asyncio]>=2.0.49",
    "alembic>=1.14",
    "asyncpg>=0.30",
    # Story 3.3 — `langgraph-checkpoint-postgres` AsyncPostgresSaver의 sync driver.
    # transitive에서 *명시 직접 의존성*으로 승격 — driver 버전·extra 통제권 + 외주
    # 인수 시 dependency tree 명확. `[binary]`은 prebuilt wheel(libpq 시스템 의존성
    # 회피, Docker `python:3.13-slim` Linux + Windows dev 동일), `[pool]`은
    # `psycopg_pool.AsyncConnectionPool` 직접 의존.
    "psycopg[binary,pool]>=3.2",
    "pgvector>=0.3",
    "pydantic-settings>=2.7",
    # --- Infra / utils ---
    "redis>=5.2",
    "httpx>=0.28",
    "tenacity>=9.0",
    "sentry-sdk>=2.20",
    "structlog>=25.1",
    "python-jose[cryptography]>=3.3",
    "passlib>=1.7",
    "slowapi>=0.1.9",
    "apscheduler>=3.11",
    "sse-starlette>=2.2",
    "google-auth>=2.36",
    "boto3>=1.36",
    # --- LLM Vision (Story 2.3 — OpenAI Vision OCR / structured outputs) ---
    "openai>=1.55",
    # --- Story 3.2 — 가이드라인 RAG 시드 + chunking 3-format 인터페이스 ---
    # text-splitters는 langchain umbrella의 별 패키지(메인 langchain의 transitive 미보장).
    "langchain-text-splitters>=0.3",
    # PDF 텍스트 추출 (BSD-3, AGPL pymupdf 회피 — D1 한글 폰트 한계는 운영자 SOP 흡수).
    "pypdf>=5.1",
    # HTML 파싱 (MIT, lxml 별 의존 X — pure Python `html.parser` 사용).
    "beautifulsoup4>=4.12",
    # YAML frontmatter 파싱 — langchain transitive일 가능성이지만 명시 등재로 직접 의존성 표기.
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "ASYNC"]
ignore = []

# Story 1.3: legal_documents.py는 한·영 약관 풀텍스트를 모듈 const로 임베드 — 본문 단락은
# 한 줄당 한 단락을 유지해 PR diff 가시성과 자연어 정확성을 우선한다(line-length 무시).
[tool.ruff.lint.per-file-ignores]
"app/domain/legal_documents.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
python_version = "3.13"
plugins = ["pydantic.mypy"]
mypy_path = "."

[[tool.mypy.overrides]]
module = ["jose.*", "passlib.*", "pgvector.*", "slowapi.*", "apscheduler.*", "google.*", "boto3.*", "botocore.*", "pypdf.*", "bs4.*", "langchain_text_splitters.*", "yaml.*", "psycopg.*", "psycopg_pool.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term --cov-fail-under=70"
testpaths = ["tests"]
markers = [
    "perf: Story 3.3 NFR-P4/P5 + T1 KPI — RUN_PERF_TESTS=1 opt-in (skip default)",
]

[tool.coverage.run]
# Story 1.2부터 security.py / exceptions.py가 채워지므로 omit 비움.
# 신규 모듈(adapters, api/deps, api/v1/auth, api/v1/users, db/models/*) 모두 측정 대상.
omit = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
